adding more

+++ src/main/java/com/example/rewards/repository/LoginRepository.java
@@ -3,6 +3,6 @@ package com.example.rewards.repository;
 import com.example.rewards.model.Login;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoginRepository extends JpaRepository<Login, Long> {
+public interface LoginRepository extends JpaRepository<Login, Integer> {
 }
